MassSpring refactor: Fix Linux warnings

+++ SurgSim/Math/Matrix.cpp
@@ -30,9 +30,9 @@ void resize<DiagonalMatrix>(DiagonalMatrix* A, unsigned int numRow, unsigned int
 	{
 		return;
 	}
-	if (A->rows() != numRow && A->cols() != numCol)
+	if (A->rows() != static_cast<int>(numRow) && A->cols() != static_cast<int>(numCol))
 	{
-		A->resize(numCol);
+		A->resize(static_cast<int>(numCol));
 	}
 	if (zeroOut)
 	{
